wip - Extracting model loading logic from model classes to ModelFactory
This is a wip commit because GPU tests still need to be fixed
