+++ deepparse/network/bpemb_seq2seq.py
@@ -13,8 +13,6 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
     BPEmb Seq2Seq network is the best of the two proposed models but takes more ``GPU``/``CPU`` resources.
 
      Args:
-        cache_dir (str): The path to the cached directory to use for downloading (and loading) the
-            model weights.
         input_size (int): The input size of the encoder (i.e. the size of the embedding). It will also be used to
             initialize the internal embeddings network input size, hidden size and output dim. The default value is
             ``300``.
@@ -26,13 +24,10 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
             ``9``.
         attention_mechanism (bool): Either or not to use the attention mechanism. The default value is ``False``.
         verbose (bool): Turn on/off the verbosity of the model. The default value is ``True`.
-        path_to_retrained_model (Union[str, None]): The path to the retrained model to use for the seq2seq.`The default
-            value is ``None``.
     """
 
     def __init__(
         self,
-        cache_dir: str,
         input_size: int = 300,
         encoder_hidden_size: int = 1024,
         encoder_num_layers: int = 1,
@@ -41,9 +36,7 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
         output_size: int = 9,
         attention_mechanism: bool = False,
         verbose: bool = True,
-        path_to_retrained_model: Union[str, None] = None,
-        pre_trained_weights: bool = True,
-        offline: bool = False,
+        version: str = "",
     ) -> None:
         super().__init__(
             input_size=input_size,
@@ -59,21 +52,6 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
         self.embedding_network = EmbeddingNetwork(
             input_size=input_size, hidden_size=input_size, projection_size=input_size
         )
-
-        model_weights_name = "bpemb"
-        if attention_mechanism:
-            model_weights_name += "_attention"
-
-        if path_to_retrained_model is not None:
-            self._load_weights(path_to_retrained_model)
-
-            version = "FineTunedModel" + self._load_version(model_type=model_weights_name, cache_dir=cache_dir)
-        elif pre_trained_weights:
-            # Means we use the pretrained weights
-            self._load_pre_trained_weights(model_weights_name, cache_dir=cache_dir, offline=offline)
-            version = self._load_version(model_type=model_weights_name, cache_dir=cache_dir)
-        else:
-            version = ""
 
         self.version = version
