+++ tests/network/test_bpemb_seq2seq_model_cpu.py
@@ -25,12 +25,11 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         cls.a_target_vector = torch.tensor([[0, 1, 1, 4, 5, 8], [1, 0, 3, 8, 0, 0]], device=cls.a_cpu_device)
         cls.a_transpose_target_vector = cls.a_target_vector.transpose(0, 1)
 
-    @patch("deepparse.network.seq2seq.Seq2SeqModel._load_pre_trained_weights")
     def test_whenInstantiatingABPEmbSeq2SeqModel_thenShouldInstantiateAEmbeddingNetwork(
-        self, load_pre_trained_weights_mock
+        self,
     ):
         seq2seq_model = BPEmbSeq2SeqModel(
-            self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+            output_size=self.output_size, verbose=self.verbose
         )
 
         self.assertEqual(self.input_size, seq2seq_model.embedding_network.model.input_size)
@@ -40,92 +39,90 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
             seq2seq_model.embedding_network.projection_layer.out_features,
         )
 
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenDefaultModel_whenLoadVersion_thenModelHash(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights"):
-            bpemb_model = BPEmbSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-
-        actual = bpemb_model.version
-        expected = self.a_bpemb_hash_value
-
-        self.assertEqual(actual, expected)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenFineTunedModel_whenLoadVersion_thenModelHashWithFineTuned(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights"):
-            bpemb_model = BPEmbSeq2SeqModel(
-                self.fake_cache_dir,
-                output_size=self.output_size,
-                verbose=self.verbose,
-                path_to_retrained_model=self.a_path_to_retrained_model,
-            )
-
-        actual = bpemb_model.version
-        expected = "FineTunedModel" + self.a_bpemb_hash_value
-
-        self.assertEqual(actual, expected)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenNotLocalWeights_whenInstantiatingABPEmbSeq2SeqModel_thenShouldDownloadWeights(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
-            BPEmbSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-            download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("deepparse.network.seq2seq.latest_version")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenLocalWeightsNotLastVersion_whenInstantiatingABPEmbSeq2SeqModel_thenShouldDownloadWeights(
-        self, load_state_dict_mock, torch_mock, isfile_mock, last_version_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = True
-        last_version_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
-            BPEmbSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-            download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("deepparse.network.seq2seq.latest_version")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenLocalWeights_whenInstantiatingABPEmbSeq2SeqModel_thenShouldntDownloadWeights(
-        self, load_state_dict_mock, torch_mock, isfile_mock, last_version_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = True
-        last_version_mock.return_value = True
-        with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
-            BPEmbSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-            download_weights_mock.assert_not_called()
-
     # TODO: MODEL LOADING LOGIC WILL BE MOVED TO MODELFACTORY !!!!!!!!!!!
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenDefaultModel_whenLoadVersion_thenModelHash(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     bpemb_model = BPEmbSeq2SeqModel(
+    #             output_size=self.output_size, verbose=self.verbose
+    #         )
+    #
+    #     actual = bpemb_model.version
+    #     expected = self.a_bpemb_hash_value
+    #
+    #     self.assertEqual(actual, expected)
+    #
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenFineTunedModel_whenLoadVersion_thenModelHashWithFineTuned(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     bpemb_model = BPEmbSeq2SeqModel(
+    #             output_size=self.output_size,
+    #             verbose=self.verbose,
+    #             path_to_retrained_model=self.a_path_to_retrained_model,
+    #         )
+    #
+    #     actual = bpemb_model.version
+    #     expected = "FineTunedModel" + self.a_bpemb_hash_value
+    #
+    #     self.assertEqual(actual, expected)
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenNotLocalWeights_whenInstantiatingABPEmbSeq2SeqModel_thenShouldDownloadWeights(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
+    #         BPEmbSeq2SeqModel(
+    #             self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+    #         )
+    #         download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
+
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenLocalWeightsNotLastVersion_whenInstantiatingABPEmbSeq2SeqModel_thenShouldDownloadWeights(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = True
+    #     last_version_mock.return_value = False
+    #     with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
+    #         BPEmbSeq2SeqModel(
+    #             self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+    #         )
+    #         download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenLocalWeights_whenInstantiatingABPEmbSeq2SeqModel_thenShouldntDownloadWeights(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = True
+    #     last_version_mock.return_value = True
+    #     with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
+    #         BPEmbSeq2SeqModel(
+    #             self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+    #         )
+    #         download_weights_mock.assert_not_called()
+
 
     # @patch("deepparse.weights_tools.torch")
     # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -148,8 +145,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Encoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -158,12 +153,10 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         encoder_mock,
         torch_load_mock,
     ):
-        seq2seq_model = BPEmbSeq2SeqModel(self.fake_cache_dir, self.output_size, self.verbose)
+        seq2seq_model = BPEmbSeq2SeqModel(self.output_size, self.verbose)
 
         to_predict_mock, lengths_list = self.setup_encoder_mocks()
         encoder_mock.__call__().return_value = (MagicMock(), MagicMock())
@@ -175,8 +168,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -185,13 +176,11 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         torch_load_mock,
     ):
         seq2seq_model = BPEmbSeq2SeqModel(
-            self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+            output_size=self.output_size, verbose=self.verbose
         )
 
         decoder_input_mock, decoder_hidden_mock = self.setUp_decoder_mocks(decoder_mock, attention_mechanism=False)
@@ -217,8 +206,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -227,13 +214,10 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         torch_load_mock,
     ):
         seq2seq_model = BPEmbSeq2SeqModel(
-            self.fake_cache_dir,
             output_size=self.output_size,
             verbose=self.verbose,
             attention_mechanism=True,
@@ -263,8 +247,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -273,8 +255,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         random_mock,
         torch_load_mock,
@@ -282,7 +262,7 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         random_mock.return_value = self.a_value_lower_than_threshold
 
         seq2seq_model = BPEmbSeq2SeqModel(
-            self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+            output_size=self.output_size, verbose=self.verbose
         )
 
         decoder_input_mock, decoder_hidden_mock = self.setUp_decoder_mocks(decoder_mock, attention_mechanism=False)
@@ -312,8 +292,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -322,8 +300,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         torch_load_mock,
@@ -377,8 +353,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -387,8 +361,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         random_mock,
@@ -443,8 +415,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -453,8 +423,6 @@ class BPEmbSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         random_mock,
