+++ tests/network/test_fasttext_seq2seq_model_cpu.py
@@ -25,76 +25,75 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         cls.a_target_vector = torch.tensor([[0, 1, 1, 4, 5, 8], [1, 0, 3, 8, 0, 0]], device=cls.a_cpu_device)
         cls.a_transpose_target_vector = cls.a_target_vector.transpose(0, 1)
 
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenDefaultModel_whenLoadVersion_thenModelHash(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights"):
-            fasttext_model = FastTextSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-
-        actual = fasttext_model.version
-        expected = self.a_fasttext_hash_value
-
-        self.assertEqual(actual, expected)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenFineTunedModel_whenLoadVersion_thenModelHashWithFineTuned(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights"):
-            fasttext_model = FastTextSeq2SeqModel(
-                self.fake_cache_dir,
-                output_size=self.output_size,
-                verbose=self.verbose,
-                path_to_retrained_model=self.a_path_to_retrained_model,
-            )
-
-        actual = fasttext_model.version
-        expected = "FineTunedModel" + self.a_fasttext_hash_value
-
-        self.assertEqual(actual, expected)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenNotLocalWeights_whenInstantiatingAFastTextSeq2SeqModel_thenShouldDownloadWeights(
-        self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
-            FastTextSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-            download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
-
-    @patch("deepparse.weights_tools.torch")
-    @patch("deepparse.network.seq2seq.latest_version")
-    @patch("os.path.isfile")
-    @patch("deepparse.network.seq2seq.torch")
-    @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
-    def test_givenLocalWeightsNotLastVersion_whenInstantiatingAFastTextSeq2SeqModel_thenShouldDownloadWeights(
-        self, load_state_dict_mock, torch_mock, isfile_mock, last_version_mock, torch_load_mock
-    ):
-        isfile_mock.return_value = True
-        last_version_mock.return_value = False
-        with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
-            FastTextSeq2SeqModel(
-                self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
-            )
-            download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
-   
     # TODO: MODEL LOADING LOGIC WILL BE MOVED TO MODELFACTORY !!!!!!!!!!!
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenDefaultModel_whenLoadVersion_thenModelHash(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     with patch("deepparse.network.seq2seq.download_weights"):
+    #         fasttext_model = FastTextSeq2SeqModel(
+    #             output_size=self.output_size, verbose=self.verbose
+    #         )
+    #
+    #     actual = fasttext_model.version
+    #     expected = self.a_fasttext_hash_value
+    #
+    #     self.assertEqual(actual, expected)
+    #
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenFineTunedModel_whenLoadVersion_thenModelHashWithFineTuned(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     with patch("deepparse.network.seq2seq.download_weights"):
+    #         fasttext_model = FastTextSeq2SeqModel(
+    #             output_size=self.output_size,
+    #             verbose=self.verbose,
+    #             path_to_retrained_model=self.a_path_to_retrained_model,
+    #         )
+    #
+    #     actual = fasttext_model.version
+    #     expected = "FineTunedModel" + self.a_fasttext_hash_value
+    #
+    #     self.assertEqual(actual, expected)
+
+
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenNotLocalWeights_whenInstantiatingAFastTextSeq2SeqModel_thenShouldDownloadWeights(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = False
+    #     with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
+    #         FastTextSeq2SeqModel(
+    #             output_size=self.output_size, verbose=self.verbose
+    #         )
+    #         download_weights_mock.assert_called_with(self.model_type, verbose=self.verbose)
+    #
+    # @patch("deepparse.weights_tools.torch")
+    # @patch("os.path.isfile")
+    # @patch("deepparse.network.seq2seq.torch")
+    # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
+    # def test_givenLocalWeightsNotLastVersion_whenInstantiatingAFastTextSeq2SeqModel_thenShouldDownloadWeights(
+    #     self, load_state_dict_mock, torch_mock, isfile_mock, torch_load_mock
+    # ):
+    #     isfile_mock.return_value = True
+    #     with patch("deepparse.network.seq2seq.download_weights") as download_weights_mock:
+    #         FastTextSeq2SeqModel(
+    #             self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+    #         )
+    #         download_weights_mock.assert_called_with(self.model_type, self.fake_cache_dir, verbose=self.verbose)
+   
     #
     # @patch("deepparse.weights_tools.torch")
     # @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -117,8 +116,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Encoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -127,12 +124,10 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         encoder_mock,
         torch_load_mock,
     ):
-        seq2seq_model = FastTextSeq2SeqModel(self.fake_cache_dir, self.output_size, self.verbose)
+        seq2seq_model = FastTextSeq2SeqModel(self.output_size, self.verbose)
 
         to_predict_mock, lengths_list = self.setup_encoder_mocks()
         encoder_mock.__call__().return_value = (MagicMock(), MagicMock())
@@ -145,8 +140,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -155,13 +148,10 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         torch_load_mock,
     ):
         seq2seq_model = FastTextSeq2SeqModel(
-            self.fake_cache_dir,
             output_size=self.output_size,
             verbose=self.verbose,
             attention_mechanism=False,
@@ -190,8 +180,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
 
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -200,13 +188,10 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         torch_load_mock,
     ):
         seq2seq_model = FastTextSeq2SeqModel(
-            self.fake_cache_dir,
             output_size=self.output_size,
             verbose=self.verbose,
             attention_mechanism=True,
@@ -236,8 +221,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -246,8 +229,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         random_mock,
         torch_load_mock,
@@ -255,7 +236,7 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         random_mock.return_value = self.a_value_lower_than_threshold
 
         seq2seq_model = FastTextSeq2SeqModel(
-            self.fake_cache_dir, output_size=self.output_size, verbose=self.verbose
+            output_size=self.output_size, verbose=self.verbose
         )
 
         decoder_input_mock, decoder_hidden_mock = self.setUp_decoder_mocks(decoder_mock, attention_mechanism=False)
@@ -285,8 +266,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.weights_tools.torch")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -295,8 +274,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         torch_load_mock,
@@ -313,7 +290,7 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         # We mock the return of the decoder output
         encoder_mock.__call__().return_value = (decoder_input_mock, decoder_hidden_mock)
 
-        seq2seq_model = FastTextSeq2SeqModel(self.fake_cache_dir, self.output_size, self.verbose)
+        seq2seq_model = FastTextSeq2SeqModel(self.output_size, self.verbose)
 
         seq2seq_model.forward(to_predict=to_predict_mock, lengths=lengths_list, target=None)
 
@@ -334,8 +311,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -344,8 +319,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         random_mock,
@@ -366,7 +339,7 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         # We mock the return of the decoder output
         encoder_mock.__call__().return_value = (decoder_input_mock, decoder_hidden_mock)
 
-        seq2seq_model = FastTextSeq2SeqModel(self.fake_cache_dir, self.output_size, self.verbose)
+        seq2seq_model = FastTextSeq2SeqModel(self.output_size, self.verbose)
 
         seq2seq_model.forward(
             to_predict=to_predict_mock,
@@ -392,8 +365,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
     @patch("deepparse.network.seq2seq.random.random")
     @patch("deepparse.network.seq2seq.Encoder")
     @patch("deepparse.network.seq2seq.Decoder")
-    @patch("deepparse.network.seq2seq.download_weights")
-    @patch("deepparse.network.seq2seq.latest_version")
     @patch("os.path.isfile")
     @patch("deepparse.network.seq2seq.torch")
     @patch("deepparse.network.seq2seq.Seq2SeqModel.load_state_dict")
@@ -402,8 +373,6 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         load_state_dict_mock,
         torch_mock,
         isfile_mock,
-        last_version_mock,
-        download_weights_mock,
         decoder_mock,
         encoder_mock,
         random_mock,
@@ -424,7 +393,7 @@ class FasttextSeq2SeqCPUTest(Seq2SeqTestCase):
         # We mock the return of the decoder output
         encoder_mock.__call__().return_value = (decoder_input_mock, decoder_hidden_mock)
         seq2seq_model = FastTextSeq2SeqModel(
-            self.fake_cache_dir, self.output_size, self.verbose, attention_mechanism=True
+            self.output_size, self.verbose, attention_mechanism=True
         )
 
         seq2seq_model.forward(
